Simplify the execution of tests
Applying the standards we have for PHP projects, where tests are simply
executed by running `phpunit` using the default configuration file
(`phpunit.xml.dist`) and maintainers/contributors can override the
default configuration using `phpunit.xml` - which is understood by
PHPUnit.

This also exposes the fact that the build was not failing even though
one of the tests was failing.